+++ internal/dnsforward/dnsforward.go
@@ -510,46 +510,6 @@ func (s *Server) Prepare(conf *ServerConfig) (err error) {
 	return nil
 }
 
-// setupDNS64 initializes DNS64 settings, the NAT64 prefixes in particular.  If
-// the DNS64 feature is enabled and no prefixes are configured, the default
-// Well-Known Prefix is used, just like Section 5.2 of RFC 6147 prescribes.
-func (s *Server) setupDNS64() (err error) {
-	if !s.conf.UseDNS64 {
-		return nil
-	}
-
-	l := len(s.conf.DNS64Prefixes)
-	if l == 0 {
-		s.dns64Prefs = []netip.Prefix{dns64WellKnownPref}
-
-		return nil
-	}
-
-	prefs := make([]netip.Prefix, 0, l)
-	for _, pref := range s.conf.DNS64Prefixes {
-		var p netip.Prefix
-		p, err = netip.ParsePrefix(pref)
-		if err != nil {
-			return err
-		}
-
-		addr := p.Addr()
-		if !addr.Is6() {
-			return fmt.Errorf("prefix %q is not an IPv6 prefix", pref)
-		}
-
-		if p.Bits() > maxNAT64PrefixBitLen {
-			return fmt.Errorf("prefix %q is too long for DNS64", pref)
-		}
-
-		prefs = append(prefs, p.Masked())
-	}
-
-	s.dns64Prefs = prefs
-
-	return nil
-}
-
 // validateBlockingMode returns an error if the blocking mode data aren't valid.
 func validateBlockingMode(mode BlockingMode, blockingIPv4, blockingIPv6 net.IP) (err error) {
 	switch mode {